+++ neoforge/src/main/java/dev/aika/smsn/api/neoforge/ModPlatformImpl.java
@@ -1,7 +1,11 @@
 package dev.aika.smsn.api.neoforge;
 
+import dev.aika.smsn.api.ModPlatform;
 import lombok.experimental.UtilityClass;
 import net.neoforged.fml.loading.FMLLoader;
+import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
+import org.apache.maven.artifact.versioning.InvalidVersionSpecificationException;
+import org.apache.maven.artifact.versioning.VersionRange;
 
 @SuppressWarnings("unused")
 @UtilityClass
@@ -13,5 +17,19 @@ public class ModPlatformImpl {
     public static String getModName(String modId) {
         if (!isModLoaded(modId)) return null;
         return FMLLoader.getLoadingModList().getModFileById(modId).moduleName();
+    }
+
+    public static ModPlatform.ModInfo getModInfo(String modId) {
+        final var info = FMLLoader.getLoadingModList().getModFileById(modId);
+        if (info == null) return null;
+        return new ModPlatform.ModInfo(modId, info.moduleName(), info.versionString());
+    }
+
+    public static boolean containsVersion(String version, String versionRange) {
+        try {
+            return VersionRange.createFromVersionSpec(versionRange).containsVersion(new DefaultArtifactVersion(version));
+        } catch (InvalidVersionSpecificationException e) {
+            return false;
+        }
     }
 }