+++ neoforge/src/main/java/dev/aika/smsn/neoforge/mixin/fancytoasts/CreditsManagerMixin_147.java
@@ -1,4 +1,4 @@
-package dev.aika.smsn.fabric.mixin.fancytoasts;
+package dev.aika.smsn.neoforge.mixin.fancytoasts;
 
 import net.bivrik.fancytoasts.core.manager.CreditsManager;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,17 +8,16 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.concurrent.CompletableFuture;
+
 @SuppressWarnings({"MixinAnnotationTarget", "RedundantSuppression"})
 @Mixin(value = CreditsManager.class, remap = false)
-public abstract class CreditsManagerMixin {
+public abstract class CreditsManagerMixin_147 {
     @Shadow
-    private CreditsManager.CreditsData cachedCredits;
+    private CreditsManager.CreditsData credits;
 
     @Shadow
-    protected abstract CreditsManager.CreditsData readCredits();
-
-    @Shadow
-    public abstract CreditsManager.CreditsData getFallback();
+    protected abstract CompletableFuture<CreditsManager.CreditsData> loadAndCombineCreditsAsync();
 
     @Inject(method = "onModInit", at = @At("HEAD"), cancellable = true)
     public void onModInit(CallbackInfo ci) {
@@ -27,8 +26,10 @@ public abstract class CreditsManagerMixin {
 
     @Inject(method = "getCredits", at = @At("HEAD"))
     public void getCredits(CallbackInfoReturnable<CreditsManager.CreditsData> cir) {
-        if (this.cachedCredits != null) return;
-        this.cachedCredits = this.readCredits();
-        if (this.cachedCredits == null) this.cachedCredits = this.getFallback();
+        if (this.credits != null) return;
+        try {
+            this.credits = this.loadAndCombineCreditsAsync().get();
+        } catch (Exception ignored) {
+        }
     }
 }