+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 **NeoForge**
 
 - Add mixins for [Alex's Caves (Unofficial Port)](https://www.curseforge.com/minecraft/mc-mods/alexs-caves-unofficial-port)
+- Add mixins for [Fancy Toasts | Better Advancements](https://www.curseforge.com/minecraft/mc-mods/fancy-toasts)
+
+**Fabric**
+
+- Add mixins for [Fancy Toasts | Better Advancements](https://www.curseforge.com/minecraft/mc-mods/fancy-toasts)
 
 ### 1.3.9