+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Unreleased
+### 1.4.1
 
 - Resolve crash caused by mixin conflict with Fancy Toasts 1.4.7 [#16](https://github.com/gizmo-ds/smsn-mod/issues/16)
 - Add optional version range check to MixinConfigPlugin